Moving the deploy script to its own folder
